backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
aiohttp==3.10.3
pydantic[email]==1.10.17
python-multipart==0.0.6
pytest==8.3.2
python-dotenv==1.0.0
bcrypt==4.1.2
SQLAlchemy==2.0.31

# 邮件发送
aiosmtplib==3.0.1
email-validator==2.1.0

# Redis（速率限制和缓存）
redis==5.0.1
slowapi==0.1.9

# 额外工具
user-agents==2.2.0
httpx==0.27.0
